Consolidate installation CI environments
Each install_ci environment build now carries every assertion for its
install path instead of rebuilding the same environment per check: the
system-Python smoke installs the core set only, the conda core import
checks ride on the conda training env, one uv -i core env covers the core
import checks, the camera probes run once on the wheel install path, and
the RL-extras wheel checks join the wheel-builder smoke.
